changed general file size handling from int to long and allow arbitrary file sizes to upload

+++ src/main/java/com/wcs/wcslib/vaadin/widget/multifileupload/component/CustomUpload.java
@@ -47,7 +47,7 @@ public class CustomUpload extends Upload implements UploadComponent {
     }
 
     @Override
-    public void setMaxFileSize(int maxFileSize) {
+    public void setMaxFileSize(long maxFileSize) {
         this.maxFileSize = maxFileSize;
     }
